feat(objectify-words): remove jsdocs and add types
