+++ src/types.ts
@@ -1,0 +1,11 @@
+export type WordI = {
+	string: string;
+	index: number;
+	wordiable: boolean;
+	color: string;
+	wordiablePos: null | number;
+};
+
+export type RegexI = {
+	wordiables: RegExp;
+};
